red_teaming_101/requirements.txt
fastapi==0.122.0
gradio==6.0.1
openai==2.8.1
python-dotenv==1.2.1
requests==2.32.5
uvicorn==0.38.0
